requirements/install.txt
# helps to get this guy first
Cython
cffi

# mine
meinconf
uninhibited
mainline

# async
aiohttp
aiorwlock
hbmqtt
kawaiisync
paco
uvloop
websockets

# sync
attrs
cachetools
#cerberus
#click
#pafy
python-mpv
rwlock
setproctitle
#sortedcontainers
six
yarl
youtube-dl

# logs
coloredlogs
colorlog

